dao: Use a long for countAll()
Use a rowCount() projection for counting. This completely get rid of
elementClass.getSimpleName() usages.

Adjust a few use cases.

+++ gemma-core/src/main/java/ubic/gemma/persistence/service/AbstractService.java
@@ -72,7 +72,7 @@ public abstract class AbstractService<O extends Identifiable> implements BaseSer
 
     @Override
     @Transactional
-    public int countAll() {
+    public long countAll() {
         return this.mainDao.countAll();
     }
